backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.3
httpx==0.26.0
supabase==2.3.0

